@@ -1,13 +1,12 @@
 {
   "name": "tiny-host-common",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "A common package for tiny hosts (auth, middleware, etc)",
   "repository": "github:MichaelFedora/tiny-host-common",
   "main": "lib/index",
   "scripts": {
     "build-lib": "del-cli lib && tsc -p tsconfig.lib.json",
-    "build-web-test": "del-cli web && webpack",
-    "build-web": "del-cli web && webpack --mode=production",
+    "build-web": "vite build",
     "start": "node scripts/host.js"
   },
   "keywords": [
@@ -19,51 +18,41 @@
   "author": "Michael Fedora",
   "license": "MIT",
   "devDependencies": {
-    "@jamescoyle/vue-icon": "^0.1.0",
-    "@mdi/js": "^5.9.55",
-    "@types/express": "^4.17.9",
-    "@types/levelup": "^4.3.0",
-    "@types/lodash": "^4.14.168",
-    "@types/node": "^14.14.35",
-    "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
-    "autoprefixer": "^10.2.5",
-    "axios": "^0.21.1",
-    "clean-webpack-plugin": "^3.0.0",
+    "@jamescoyle/vue-icon": "^0.1.2",
+    "@mdi/js": "^6.5.95",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/levelup": "^4.3.3",
+    "@types/lodash": "^4.14.176",
+    "@types/node": "^16.11.7",
+    "@types/uuid": "^8.3.1",
+    "@typescript-eslint/eslint-plugin": "^5.3.1",
+    "@typescript-eslint/parser": "^5.3.1",
+    "@vitejs/plugin-vue": "^1.9.4",
+    "@vue/compiler-sfc": "^3.2.21",
+    "autoprefixer": "^10.4.0",
+    "axios": "^0.24.0",
     "cors": "^2.8.5",
-    "css-loader": "^5.1.3",
-    "del-cli": "^3.0.1",
-    "encoding-down": "^6.3.0",
-    "eslint": "^7.22.0",
-    "eslint-plugin-vue": "^7.8.0",
-    "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.0",
-    "html-webpack-plugin": "^5.3.1",
+    "del-cli": "^4.0.1",
+    "encoding-down": "^7.1.0",
+    "eslint": "^8.2.0",
+    "eslint-plugin-vue": "^8.0.3",
     "lodash": "^4.17.21",
-    "memdown": "^5.1.0",
-    "mini-css-extract-plugin": "^1.3.9",
-    "node-sass": "^5.0.0",
+    "memdown": "^6.1.1",
     "normalize.css": "^8.0.1",
-    "postcss": "^8.2.8",
-    "postcss-loader": "^5.2.0",
-    "sass-loader": "^11.0.1",
-    "terser-webpack-plugin": "^5.1.1",
-    "ts-loader": "^8.0.18",
-    "tsconfig-paths-webpack-plugin": "^3.5.1",
-    "typescript": "^4.2.3",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.6",
-    "vue-router": "^3.5.1",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.27.2",
-    "webpack-cli": "^4.5.0"
+    "postcss": "^8.3.11",
+    "sass": "^1.43.4",
+    "terser": "^5.9.0",
+    "typescript": "^4.4.4",
+    "vite": "^2.6.14",
+    "vue": "^3.2.21",
+    "vue-router": "^4.0.12"
   },
   "dependencies": {
     "express": "^4.17.1",
-    "levelup": "^4.4.0",
-    "rxjs": "^6.6.6",
-    "source-map-support": "^0.5.19",
+    "levelup": "^5.1.1",
+    "rxjs": "^7.4.0",
+    "source-map-support": "^0.5.20",
     "uuid": "^8.3.2"
   }
 }
