@@ -8,54 +8,54 @@ const middleware_1 = require("../middleware");
 const auth_middleware_1 = require("./auth-middleware");
 const auth_util_1 = require("./auth-util");
 class AuthApi {
-    constructor(config, db, router = express_1.Router()) {
+    constructor(config, db, router = (0, express_1.Router)()) {
         config = Object.assign({ requireScopes: true, allowHandshakes: true, allowMasterKeys: true, handshakeExpTime: 300000 }, config);
         this._router = router;
-        const validateSession = auth_middleware_1.validateUserSession(db);
+        const validateSession = (0, auth_middleware_1.validateUserSession)(db);
         router.get('/type', (_, res) => res.send(''));
         // auth
-        const authRouter = express_1.Router();
-        authRouter.post('/login', express_1.json(), middleware_1.wrapAsync(async (req, res) => {
+        const authRouter = (0, express_1.Router)();
+        authRouter.post('/login', (0, express_1.json)(), (0, middleware_1.wrapAsync)(async (req, res) => {
             if (config.whitelist && !config.whitelist.includes(req.body.username))
                 throw new errors_1.AuthError('Whitelist is active.');
             const user = await db.getUserFromUsername(req.body.username);
             if (!user)
                 throw new errors_1.AuthError('Username / password mismatch.');
-            const pass = await auth_util_1.hash(user.salt, req.body.password);
+            const pass = await (0, auth_util_1.hash)(user.salt, req.body.password);
             if (user.pass !== pass)
                 throw new errors_1.AuthError('Username / password mismatch.');
             res.send(await db.addSession(user.id, ['!user', '/']));
         }), middleware_1.handleValidationError);
-        authRouter.post('/register', express_1.json(), middleware_1.wrapAsync(async (req, res) => {
+        authRouter.post('/register', (0, express_1.json)(), (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.body.username || !req.body.password)
                 throw new errors_1.MalformedError('Must have a username and password!');
             if (config.whitelist && !config.whitelist.includes(req.body.username))
                 throw new errors_1.NotAllowedError('Whitelist is active.');
             if (await db.getUserFromUsername(req.body.username))
                 throw new errors_1.NotAllowedError('Username taken!');
-            const salt = crypto_1.randomBytes(128).toString('hex');
+            const salt = (0, crypto_1.randomBytes)(128).toString('hex');
             const user = {
                 username: req.body.username,
                 salt,
-                pass: await auth_util_1.hash(salt, req.body.password)
+                pass: await (0, auth_util_1.hash)(salt, req.body.password)
             };
             await db.addUser(user);
             res.sendStatus(204);
         }));
-        authRouter.post('/change-pass', validateSession, express_1.json(), middleware_1.wrapAsync(async (req, res) => {
+        authRouter.post('/change-pass', validateSession, (0, express_1.json)(), (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.session.scopes.includes('!user'))
                 throw new errors_1.NotAllowedError('Must be a user!');
             if (!req.body.password || !req.body.newpass)
                 throw new errors_1.MalformedError('Body must have a password, and a newpass.');
-            if (await auth_util_1.hash(req.user.salt, req.body.password) !== req.user.pass)
+            if (await (0, auth_util_1.hash)(req.user.salt, req.body.password) !== req.user.pass)
                 throw new errors_1.NotAllowedError('Password mismatch.');
-            const salt = crypto_1.randomBytes(128).toString('hex');
-            const pass = await auth_util_1.hash(salt, req.body.newpass);
+            const salt = (0, crypto_1.randomBytes)(128).toString('hex');
+            const pass = await (0, auth_util_1.hash)(salt, req.body.newpass);
             await db.putUser(req.user.id, Object.assign(req.user, { salt, pass }));
             const sessions = await db.getSessionIdsForUser(req.user.id);
             await db.delManySessions(sessions.filter(a => a !== req.session.id));
         }));
-        authRouter.get('/sessions', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        authRouter.get('/sessions', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.session.scopes.includes('!user'))
                 throw new errors_1.NotAllowedError('Must be a user!');
             const sessions = await db.getSessionsForUser(req.user.id);
@@ -63,7 +63,7 @@ class AuthApi {
                 delete sess.user;
             res.json(sessions);
         }));
-        authRouter.delete('/sessions/:id', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        authRouter.delete('/sessions/:id', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.session.scopes.includes('!user'))
                 throw new errors_1.NotAllowedError('Must be a user!');
             const sess = await db.getSession(req.params.id);
@@ -72,25 +72,25 @@ class AuthApi {
             await db.delSession(sess.id);
             res.sendStatus(204);
         }));
-        authRouter.delete('/sessions', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        authRouter.delete('/sessions', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.session.scopes.includes('!user'))
                 throw new errors_1.NotAllowedError('Must be a user!');
             const sessions = await db.getSessionIdsForUser(req.params.id);
             await db.delManySessions(sessions.filter(a => a !== req.session.id));
             res.sendStatus(204);
         }));
-        authRouter.post('/logout', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        authRouter.post('/logout', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             await db.delSession(req.session.id);
             res.sendStatus(204);
         }));
-        authRouter.get('/refresh', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        authRouter.get('/refresh', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             const sess = await db.addSession(req.user.id, req.session.scopes);
             await db.delSession(req.session.id);
             res.json(sess);
         }));
         if (config.allowHandshakes) {
-            const handshakeRouter = express_1.Router();
-            handshakeRouter.get('/start', middleware_1.wrapAsync(async (req, res) => {
+            const handshakeRouter = (0, express_1.Router)();
+            handshakeRouter.get('/start', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.query.redirect || typeof req.query.redirect !== 'string' ||
                     (req.query.scopes ? typeof req.query.scopes !== 'string' : config.requireScopes))
                     throw new errors_1.MalformedError('Must have ?redirect={url}<&scopes=["/scopes"]> query.');
@@ -107,7 +107,7 @@ class AuthApi {
                 const hsId = await db.addHandshake(info);
                 res.redirect(`/handshake?handshake=${hsId}${req.query.username ? '&username=' + String(req.query.username) : ''}`);
             }));
-            handshakeRouter.use('/:id', validateSession, middleware_1.wrapAsync(async (req, res, next) => {
+            handshakeRouter.use('/:id', validateSession, (0, middleware_1.wrapAsync)(async (req, res, next) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 req.handshake = await db.getHandshake(req.params.id);
@@ -119,7 +119,7 @@ class AuthApi {
                 }
                 next();
             }));
-            handshakeRouter.get('/:id', middleware_1.wrapAsync(async (req, res) => {
+            handshakeRouter.get('/:id', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 res.json({
@@ -127,12 +127,12 @@ class AuthApi {
                     scopes: req.handshake.scopes
                 });
             }));
-            handshakeRouter.get('/:id/approve', middleware_1.wrapAsync(async (req, res) => {
+            handshakeRouter.get('/:id/approve', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 let code;
                 do {
-                    code = crypto_1.randomBytes(24).toString('hex');
+                    code = (0, crypto_1.randomBytes)(24).toString('hex');
                 } while (await db.getHandshakeFromCode(code) != null);
                 const info = {
                     user: req.user.id,
@@ -141,14 +141,14 @@ class AuthApi {
                 await db.putHandshake(req.handshake.id, Object.assign(req.handshake, info));
                 res.redirect(req.handshake.redirect + '?code=' + code);
             }));
-            handshakeRouter.get('/:id/cancel', middleware_1.wrapAsync(async (req, res) => {
+            handshakeRouter.get('/:id/cancel', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 await db.delHandshake(req.handshake.id);
                 res.redirect(req.handshake.redirect + '?error=access_denied');
             }));
             authRouter.use('/handshake', handshakeRouter);
-            authRouter.post('/session', middleware_1.wrapAsync(async (req, res) => {
+            authRouter.post('/session', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.body.redirect || !(!config.requireScopes || req.body.scopes) || !req.body.code)
                     throw new errors_1.MalformedError('Body should contain: { app, redirect, scopes, code }!');
                 const handshake = await db.getHandshakeFromCode(req.body.code);
@@ -165,19 +165,19 @@ class AuthApi {
             }));
         }
         if (config.allowMasterKeys) {
-            const masterKeyRouter = express_1.Router();
-            masterKeyRouter.get('', validateSession, middleware_1.wrapAsync(async (req, res) => {
+            const masterKeyRouter = (0, express_1.Router)();
+            masterKeyRouter.get('', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 res.json(await db.getMasterKeysForUser(req.user.id));
             }));
-            masterKeyRouter.post('', validateSession, middleware_1.wrapAsync(async (req, res) => {
+            masterKeyRouter.post('', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 const name = ((req.query.name && typeof req.query.name === 'string') ? String(req.query.name) : '') || 'Unknown';
                 res.json(await db.addMasterKey({ user: req.user.id, name }));
             }));
-            masterKeyRouter.put('/:id', validateSession, express_1.json(), middleware_1.wrapAsync(async (req, res) => {
+            masterKeyRouter.put('/:id', validateSession, (0, express_1.json)(), (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 if (!req.body || !req.body.name || typeof req.body.name !== 'string')
@@ -188,7 +188,7 @@ class AuthApi {
                 await db.putMasterKey(req.params.id, Object.assign({}, masterkey, { name: req.body.name }));
                 res.sendStatus(204);
             }));
-            masterKeyRouter.delete('/:id', validateSession, middleware_1.wrapAsync(async (req, res) => {
+            masterKeyRouter.delete('/:id', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.session.scopes.includes('!user'))
                     throw new errors_1.NotAllowedError('Must be a user!');
                 const masterkey = await db.getMasterKey(req.params.id);
@@ -197,8 +197,8 @@ class AuthApi {
                 await db.delMasterKey(masterkey.id);
                 res.sendStatus(204);
             }));
-            authRouter.use('/master-key', masterKeyRouter, middleware_1.handleError('auth-master-key'));
-            authRouter.post('/generate-session', middleware_1.wrapAsync(async (req, res) => {
+            authRouter.use('/master-key', masterKeyRouter, (0, middleware_1.handleError)('auth-master-key'));
+            authRouter.post('/generate-session', (0, middleware_1.wrapAsync)(async (req, res) => {
                 if (!req.query.key || typeof req.query.key !== 'string')
                     throw new errors_1.MalformedError('?key=.. required!');
                 if (req.query.scopes ? typeof req.query.scopes !== 'string' : config.requireScopes)
@@ -215,20 +215,20 @@ class AuthApi {
                 res.json(await db.addSession(user.id, scopes));
             }));
         }
-        router.use('/auth', authRouter, middleware_1.handleError('auth'));
-        router.get('/self', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        router.use('/auth', authRouter, (0, middleware_1.handleError)('auth'));
+        router.get('/self', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             res.json({ id: req.user.id, username: req.user.username });
-        }), middleware_1.handleError('get-self'));
-        router.delete('/self', validateSession, middleware_1.wrapAsync(async (req, res) => {
+        }), (0, middleware_1.handleError)('get-self'));
+        router.delete('/self', validateSession, (0, middleware_1.wrapAsync)(async (req, res) => {
             if (!req.session.scopes.includes('!user'))
                 throw new errors_1.NotAllowedError('Must be a user!');
             if (!req.query.pass || typeof req.query.pass !== 'string')
                 throw new errors_1.MalformedError('?pass=.. required!');
-            if (await auth_util_1.hash(req.user.salt, req.query.pass) !== req.user.pass)
+            if (await (0, auth_util_1.hash)(req.user.salt, req.query.pass) !== req.user.pass)
                 throw new errors_1.NotAllowedError('Password does not match!');
             await db.delUser(req.user.id);
             res.sendStatus(204);
-        }), middleware_1.handleError('delete-self'));
+        }), (0, middleware_1.handleError)('delete-self'));
     }
     get router() { return this._router; }
     validateScopes(scopes, requireScopes = false) {
