@@ -36,7 +36,7 @@ class AuthDB {
     async addSession(user, scopes = []) {
         let id;
         do {
-            id = uuid_1.v4();
+            id = (0, uuid_1.v4)();
         } while (await this.getSession(id) != null);
         await this.db.put(this.scope + 'session!!' + id, { user, created: Date.now(), scopes });
         return id;
@@ -100,7 +100,7 @@ class AuthDB {
         delete user.id;
         let id;
         do {
-            id = uuid_1.v4();
+            id = (0, uuid_1.v4)();
         } while (await this.getUser(id) != null);
         await this.db.put(this.scope + 'user!!' + id, user);
         return id;
@@ -142,7 +142,7 @@ class AuthDB {
         delete hs.id;
         let id;
         do {
-            id = uuid_1.v4();
+            id = (0, uuid_1.v4)();
         } while (await this.getHandshake(id) != null);
         delete hs.code;
         delete hs.user;
@@ -198,7 +198,7 @@ class AuthDB {
         delete key.id;
         let id;
         do {
-            id = uuid_1.v4();
+            id = (0, uuid_1.v4)();
         } while (await this.getMasterKey(id) != null);
         await this.db.put(this.scope + 'masterkey!!' + id, key);
         return id;
