@@ -4,7 +4,7 @@ exports.validateUserSession = void 0;
 const errors_1 = require("../errors");
 const middleware_1 = require("../middleware");
 function validateUserSession(db) {
-    return middleware_1.wrapAsync(async function (req, res, next) {
+    return (0, middleware_1.wrapAsync)(async function (req, res, next) {
         try {
             const session = await db.getSession(String(req.query.sid || '') || '');
             if (!session)
@@ -16,7 +16,7 @@ function validateUserSession(db) {
             req.user = user;
         }
         catch (e) {
-            return middleware_1.handleValidationError(e, req, res, next);
+            return (0, middleware_1.handleValidationError)(e, req, res, next);
         }
         next();
     });
