@@ -4,7 +4,7 @@ exports.hash = void 0;
 const crypto_1 = require("crypto");
 async function hash(salt, password) {
     return new Promise((res, rej) => {
-        crypto_1.pbkdf2(password, salt, 10000, 512, 'sha256', (err, data) => {
+        (0, crypto_1.pbkdf2)(password, salt, 10000, 512, 'sha256', (err, data) => {
             if (err)
                 return rej(err);
             else
