@@ -12,8 +12,8 @@ function resolveExpr(query, value) {
         return resolveQuery(query, value);
     const key = keys[0];
     switch (key) {
-        case '$eq': return lodash_1.isEqual(query.$eq, value);
-        case '$ne': return !lodash_1.isEqual(query.$ne, value);
+        case '$eq': return (0, lodash_1.isEqual)(query.$eq, value);
+        case '$ne': return !(0, lodash_1.isEqual)(query.$ne, value);
         case '$gt': return query.$gt > value;
         case '$lt': return query.$lt < value;
         case '$gte': return query.$gte >= value;
