@@ -1,7 +1,7 @@
 <template>
 <div id='tiny-loading'></div>
 </template>
-<script>
+<script lang='ts'>
 export default {
   name: 'tiny-loading'
 }
