@@ -1,30 +1,21 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
+import { createRouter, createWebHistory } from 'vue-router';
 
-// @ts-ignore
-const HomePage = () => import('pages/home.vue');
-// @ts-ignore
-const SessionsPage = () => import('pages/sessions.vue');
-// @ts-ignore
-const LoginPage = () => import('pages/login.vue');
-// @ts-ignore
-const HandshakePage = () => import('pages/handshake.vue');
+const HomePage = () => import('@/pages/home.vue');
+const SessionsPage = () => import('@/pages/sessions.vue');
+const LoginPage = () => import('@/pages/login.vue');
+const HandshakePage = () => import('@/pages/handshake.vue');
 
-// @ts-ignore
-import NotFoundPage from 'pages/not-found.vue';
-import dataBus from 'services/data-bus';
+import NotFoundPage from '@/pages/not-found.vue';
+import dataBus from '@/services/data-bus';
 
-
-Vue.use(VueRouter);
-
-const router = new VueRouter({
-  mode: 'history',
+const router = createRouter({
+  history: createWebHistory(),
   routes: [
     { path: '/', component: HomePage, name: 'home' }, // main "menu"
     { path: '/sessions', component: SessionsPage, name: 'sessions' }, // managing sessions
     { path: '/login', component: LoginPage, name: 'login' }, // loggin in (obv)
     { path: '/handshake', component: HandshakePage, name: 'handshake' }, // handshaking apps
-    { path: '**', component: NotFoundPage, name: 'not-found' }
+    { path: '/:pathMatch(.*)*', component: NotFoundPage, name: 'not-found' }
   ]
 });
 
@@ -33,7 +24,7 @@ router.beforeEach((to, from, next) => {
 
   if(to.path !== from.path) {
     if(to.name)
-      document.title = baseTitle + ' - ' + to.name;
+      document.title = baseTitle + ' - ' + to.name.toString().split('-').map(v => v[0].toLocaleUpperCase() + v.slice(1)).join(' ');
     else if(to.path.length > 1) {
 
       const sdir = [];
